Implement a flexible jump utility for finding a verse or word identifier
I chose not to validate the chapter/verse numbers -- you can jump to
Revelation 400 and it won't complain. If that proves to be problematic
we could start handling it. It could be helpful if you enter an
abbreviation with a chapter/verse number because we could eliminate any
book that doesn't have that chapter/verse. E.g. P 119 is likely to be
Psalms 119 and unlikely to be Philemon 119 since there is no Philemon
119.